Remover arquivos de frontend da pasta de backend
