--- backend/client/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-
-    <title>Quotes</title>
-    <link rel="stylesheet" href="mystyle.css">
-    <script type="text/javascript" src="quotes.js"></script>
-</head>
-
-<body>
-  <div>
-    <div class="flex-container">
-        <blockquote>
-          <span id="text"></span>
-            <i><p>
-            <span id="author"></span>
-            </p></i>
-        </blockquote>
-    </div>
-    <div class="flex-button">
-      <button class="btn" id="newQuote">&#8635;</button>
-    </div>
-  </div>
-</body>
-</html>
