--- backend/client/mystyle.css
@@ -1,63 +1,0 @@
-html, body {
-  height: 100%;
-}
-
-body {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: static;
-}
-
-.flex-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.flex-button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  outline: none;
-}
-
-blockquote {
-  min-height: 5em;
-  padding: 1em 4em;
-  font: 1em/150% sans-serif;
-  position: relative;
-  background-color: lightblue;
-}
-
-blockquote::before,
-blockquote::after {
-  position: absolute;
-  height: 3rem;
-  font: 6rem/100% Georgia, "Times New Roman", Times, serif;
-}
-
-blockquote::before {
-  content: "“";
-  top: 0.3rem;
-  left: 0.9rem;
-}
-
-blockquote::after {
-  content: "”";
-  bottom: 0.3rem;
-  right: 0.8rem;
-}
-
-blockquote i {
-  display: block;
-  font-size: 0.8em;
-  margin-top: 1rem;
-  text-align: right;
-}
-
-.btn {
-  border: none;
-  background-color: white;
-  font-size: 2rem;
-}
