--- backend/routes/index.js
@@ -1,6 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-
-
-module.exports = router;
